devtools/client/webide/test/test_fullscreenToolbox.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript" src="head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>

    <script type="application/javascript">
      function connectToLocal(win, docRuntime) {
        return new Promise(resolve => {
          win.AppManager.connection.once(
              win.Connection.Events.CONNECTED,
              resolve);
          docRuntime.querySelectorAll(".runtime-panel-item-other")[1].click();
        });
      }

      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        Task.spawn(function* () {
          let win = yield openWebIDE();
          let docProject = getProjectDocument(win);
          let docRuntime = getRuntimeDocument(win);
          win.AppManager.update("runtime-list");

          connectToLocal(win, docRuntime);

          // Select main process
          yield waitForUpdate(win, "runtime-targets");
          SimpleTest.executeSoon(() => {
            docProject.querySelectorAll("#project-panel-runtimeapps .panel-item")[0].click();
          });

          yield waitForUpdate(win, "project");

          ok(win.UI.toolboxPromise, "Toolbox promise exists");
          yield win.UI.toolboxPromise;

          let nbox = win.document.querySelector("#notificationbox");
          ok(!nbox.hasAttribute("toolboxfullscreen"), "Toolbox is not fullscreen");

          win.Cmds.showRuntimeDetails();

          ok(!nbox.hasAttribute("toolboxfullscreen"), "Toolbox is not fullscreen");

          yield win.Cmds.disconnectRuntime();

          yield closeWebIDE(win);

          DebuggerServer.destroy();

          SimpleTest.finish();
        });
      }
    </script>
  </body>
</html>
